prepare refactoring for anonymizer with JSON files
