+++ iidm/iidm-converter-api/src/main/java/com/powsybl/iidm/anonymizer/Anonymizer.java
@@ -4,7 +4,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.iidm.xml;
+package com.powsybl.iidm.anonymizer;
 
 import com.powsybl.iidm.network.Country;
