+++ iidm/iidm-converter-api/src/main/java/com/powsybl/iidm/anonymizer/ConverterContext.java
@@ -4,16 +4,16 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package com.powsybl.iidm.xml;
+package com.powsybl.iidm.anonymizer;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class XmlContext {
+public class ConverterContext {
 
     private final Anonymizer anonymizer;
 
-    public XmlContext(Anonymizer anonymizer) {
+    public ConverterContext(Anonymizer anonymizer) {
         this.anonymizer = anonymizer != null ? anonymizer : new FakeAnonymizer();
     }
